Move outputs/ to inouts/
Not sure the name won't change in the future, but it surely is more
appropriate.
